Add Luau compatibility lua.hpp header file
This PR adds a compatibility layer for standard Lua's lua.hpp in Luau. This addresses another build error:
- Created source/cpp/luau/lua.hpp to provide a C++ compatibility header
- This allows source/library.cpp to properly include <lua.hpp> at build time

Combined with the previous fix for "unsupported Lua version" in lfs.c, this should address the major Lua-related build errors.

Note that there are still other build issues with:
1. Missing JNI headers (Java Native Interface)
2. UIWebView deprecation
3. C++ standard library issues (setw, setfill)
4. Objective-C++ block variables

These will need to be addressed in subsequent PRs.