Fix CMake binary directory conflict
Fixed the CMake error where the build was failing with:
"The binary directory is already used to build a source directory"

The solution is to specify a unique binary directory name for the VM build, which prevents conflicts when CMake tries to build multiple source directories. Changed:

```cmake
add_subdirectory(VM)
```

to:

```cmake
add_subdirectory(VM VM_build)
```

This tells CMake to use "VM_build" as the binary directory for the VM source directory, which should prevent the conflict during the build process.